split out some functions into a new file and rearranged others in client.go
